requirements.txt
-e .
aws-cdk.core
aws-cdk.aws_ec2
aws-cdk.aws_iam
aws-cdk.aws_globalaccelerator
aws-cdk.aws_elasticloadbalancingv2
aws_cdk.aws_elasticloadbalancingv2-targets
aws_cdk.aws_route53
aws_cdk.aws_route53_targets
aws_cdk.aws_certificatemanager